exception not found and bad request

+++ src/main/java/com/simpla/project/services/UserService.java
@@ -4,10 +4,12 @@ import java.util.List;
 import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.stereotype.Service;
 
 import com.simpla.project.entities.User;
 import com.simpla.project.repositories.UserRepository;
+import com.simpla.project.services.exceptions.DatabaseException;
 import com.simpla.project.services.exceptions.ResourceNotFoundException;
 
 /* Classe de serviço responsável por conter a lógica de negócio relacionada à entidade User.
@@ -32,8 +34,15 @@ public class UserService {
 		return repository.save(obj);
 	}
 	
-	public void delete(Long id) {
-		repository.deleteById(id);
+	public void delete(Long id){
+	    try {
+	        if(!repository.existsById(id)) throw new ResourceNotFoundException(id);
+	        repository.deleteById(id);
+	    }catch (ResourceNotFoundException e){
+	        throw new ResourceNotFoundException(id);
+	    }catch (DataIntegrityViolationException e) {
+	    	throw new DatabaseException(e.getMessage());
+	    }
 	}
 	
 	public User update(Long id, User obj) {
